JAVA-35559: Supress spring-boot-testing-3 postgres integration test failures (#16796)
