+++ spring-boot-modules/spring-boot-testing-3/pom.xml
@@ -73,6 +73,7 @@
     </build>
 
     <properties>
+        <java.version>17</java.version>
         <start-class>com.baeldung.Application</start-class>
         <gson.version>2.10.1</gson.version>
     </properties>
